Move local server query route from hub engine to hub lib
