+++ core/engines/hub/package.json
@@ -5,7 +5,6 @@
   "client": "client.js",
   "server": "server.js",
   "dependencies": {
-    "dataurl-to-blob": "0.0.1",
-    "ps-node": "^0.1.5"
+    "dataurl-to-blob": "0.0.1"
   }
 }
