+++ lib/hub.js
@@ -6,6 +6,7 @@ const mkdirp = require('mkdirp');
 const bodyParser = require('body-parser');
 const bodyParserJson = bodyParser.json();
 const LRU = require('lru-cache');
+const psNode = require('ps-node');
 const passwordHash = require('password-hash-and-salt');
 const cookie = require('cookie');
 const Busboy = require('busboy');
@@ -446,9 +447,85 @@ const listen = (a, config, {key, userDb}) => {
         connection.close();
       }
     };
+    const _listenLocalRequest = () => {
+      a.app.get('/servers/local.json', _filterIsHubHostname, (req, res, next) => {
+        psNode.lookup({
+          command:'node',
+        }, (err, processes) => {
+          if (!err) {
+            Promise.all(processes.map(process => new Promise((accept, reject) => {
+              const {pid} = process;
+
+              fs.readlink(path.join('/', 'proc', pid, 'cwd'), (err, cwdPath) => {
+                if (!err) {
+                  const {arguments: args} = process;
+                  const binPath = args[0];
+                  const processIndexFilePath = _resolvePath(binPath, cwdPath);
+
+                  const _findArg = name => {
+                    for (let i = 0; i < args.length; i++) {
+                      const arg = args[i];
+                      const match = arg.match(/^(.+?)=(.+?)$/);
+
+                      if (match && match[1] === name) {
+                        return match[2];
+                      }
+                    }
+                    return null;
+                  };
+
+                  if (processIndexFilePath === localIndexFilePath) {
+                    const dataDirectory = _findArg('dataDirectory');
+
+                    if (dataDirectory) { // XXX check that this hub actually owns the data directory
+                      const worldname = _findArg('worldname') || null;
+                      const url = (() => {
+                        const serverHost = _findArg('serverHost');
+                        const port = _findArg('port');
+
+                        if (serverHost && port) {
+                          return serverHost + ':' + port;
+                        } else {
+                          return null;
+                        }
+                      })();
+                      const users = []; // XXX report a real users list
+                      const serverDataDirectory = _resolvePath(dataDirectory, cwdPath); // XXX no need to report this
+
+                      accept({
+                        worldname: worldname,
+                        url: url,
+                        users: users,
+                        dataDirectory: serverDataDirectory,
+                      });
+                    } else {
+                      accept(null);
+                    }
+                  } else {
+                    accept(null);
+                  }
+                } else {
+                  reject(err);
+                }
+              });
+            })))
+              .then(servers => servers.filter(server => server !== null))
+              .then(servers => {
+                res.json({
+                  servers,
+                });
+              });
+          } else {
+            res.status(500);
+            res.send(err.stack);
+          }
+        });
+      });
+    };
 
     _listenServersRequest();
     _listenServersWs();
+    _listenLocalRequest();
 
     accept();
   });
